The skip command can now skip to a song using the queue item number
Is there a song you really like deep in the queue? Using the queue command will give you a list of song numbers, then you may use the skip command followed by the queue item number to bring it to the number one spot in the queue and to discard the current playing song for that one!